+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v3.0.1] - 2023-10-02
+
+### Changed
+
+- Service provider will look for and use `false` credentials if it has been configured for the SDK or the specific
+  service. This allows for disabling credentials in testing contexts, or using fake AWS service providers such as
+  Localstack and DynamoDB Local.
+
 ## [v3.0.0] - 2023-09-07
 
 ### Added
@@ -29,5 +37,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - The configuration for the Laravel AWS SDK service provider is now passed to the AWS SDK credential provider. This
   fixes the issue of the STS client always using the `us-east-1` regional endpoint.
 
-[v3.0.0]: https://github.com/HealthEngineAU/laravel-easy-aws/compare/v2.1.2...v3.0.0
+[v3.0.1]: https://github.com/HealthEngineAU/laravel-easy-aws/compare/v3.0.0...v3.0.1
+[v3.0.0]: https://github.com/HealthEngineAU/laravel-easy-aws/compare/v2.1.1...v3.0.0
 [v2.1.1]: https://github.com/HealthEngineAU/laravel-easy-aws/compare/v2.1.0...v2.1.1